fix(cargo-tangle): CLI bugs
Closes #404
Closes #405

+++ blueprint-manager/src/config.rs
@@ -1,31 +1,31 @@
+use clap::Parser;
 use std::path::PathBuf;
-use structopt::StructOpt;
 
-#[derive(Debug, StructOpt)]
-#[structopt(
+#[derive(Debug, Parser)]
+#[command(
     name = "Blueprint Manager",
     about = "An program executor that connects to the Tangle network and runs protocols dynamically on the fly"
 )]
 pub struct BlueprintManagerConfig {
     /// The path to the gadget configuration file
-    #[structopt(parse(from_os_str), short = "s", long)]
+    #[arg(short = 's', long)]
     pub gadget_config: Option<PathBuf>,
     /// The path to the keystore
-    #[structopt(short = "k", long)]
+    #[arg(short = 'k', long)]
     pub keystore_uri: String,
     /// The directory in which all gadgets will store their data
-    #[structopt(long, short = "d", parse(from_os_str), default_value = "./data")]
+    #[arg(long, short = 'd', default_value = "./data")]
     pub data_dir: PathBuf,
     /// The verbosity level, can be used multiple times
-    #[structopt(long, short = "v", parse(from_occurrences))]
-    pub verbose: i32,
+    #[arg(long, short = 'v', action = clap::ArgAction::Count)]
+    pub verbose: u8,
     /// Whether to use pretty logging
-    #[structopt(long)]
+    #[arg(long)]
     pub pretty: bool,
     /// An optional unique string identifier for the blueprint manager to differentiate between multiple
     /// running instances of a BlueprintManager (mostly for debugging purposes)
-    #[structopt(long, short = "id")]
+    #[arg(long, alias = "id")]
     pub instance_id: Option<String>,
-    #[structopt(long, short = "t")]
+    #[arg(long, short = 't')]
     pub test_mode: bool,
 }
